Readd Xwt.Gtk.dll.
I'm suggesting that we readd this because it is a public package on
nuget.org. I'm largely assuming that no one outside of Xamarin is using
it but since it was there and public we'd be breaking existing
functionality.

Maybe we could add some kind of warning if someone tries to use it?